requirements.txt
arabic-reshaper==2.1.3
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cycler==0.11.0
fonttools==4.28.1
future==0.18.2
hazm==0.7.0
idna==3.3
igraph==0.9.8
kiwisolver==1.3.2
louvain==0.7.1
matplotlib==3.5.0
nltk==3.3
numpy==1.21.4
packaging==21.2
pandas==1.4.0
Pillow==8.4.0
pycairo==1.20.1
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.3
pywin32==303
requests==2.26.0
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
soupsieve==2.2.1
texttable==1.6.4
tomli==1.2.2
urllib3==1.26.7
XlsxWriter==3.0.2
xlwings==0.25.3
